SONARPY-1023 Move methods to add quickfixes outside of public API
