+++ python-frontend/src/main/java/org/sonar/plugins/python/api/PythonCheck.java
@@ -37,8 +37,6 @@ public interface PythonCheck {
     private final IssueLocation primaryLocation;
     private Integer cost;
     private final List<IssueLocation> secondaryLocations;
-    // TODO NEED TO HIDE, should not be in the public api for now
-    private final List<PythonQuickFix> quickFixes = new ArrayList<>();
 
     public PreciseIssue(PythonCheck check, IssueLocation primaryLocation) {
       this.check = check;
@@ -82,14 +80,6 @@ public interface PythonCheck {
 
     public List<IssueLocation> secondaryLocations() {
       return secondaryLocations;
-    }
-
-    public void addQuickFix(PythonQuickFix quickFix){
-      this.quickFixes.add(quickFix);
-    }
-
-    public List<PythonQuickFix> getQuickFixes() {
-      return quickFixes;
     }
 
     public PythonCheck check() {
