+++ python-frontend/src/main/java/org/sonar/python/quickfix/IssueWithQuickFix.java
@@ -1,0 +1,44 @@
+/*
+ * SonarQube Python Plugin
+ * Copyright (C) 2011-2022 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package org.sonar.python.quickfix;
+
+import java.util.ArrayList;
+import java.util.List;
+import org.sonar.plugins.python.api.IssueLocation;
+import org.sonar.plugins.python.api.PythonCheck;
+import org.sonar.python.reporting.PythonQuickFix;
+
+public class IssueWithQuickFix extends PythonCheck.PreciseIssue {
+
+  private final List<PythonQuickFix> quickFixes = new ArrayList<>();
+
+  public IssueWithQuickFix(PythonCheck check, IssueLocation primaryLocation) {
+    super(check, primaryLocation);
+  }
+
+  public void addQuickFix(PythonQuickFix quickFix){
+    this.quickFixes.add(quickFix);
+  }
+
+  public List<PythonQuickFix> getQuickFixes() {
+    return quickFixes;
+  }
+
+}
